chore(THU-369): lazy-load OIDC redirect and clean up tests
- extract OidcRedirect into its own lazy-loaded module so non-OIDC
  deployments skip the extra bundle weight
- rename VITE_AUTH_MODE value from 'consumer' to 'thunderbolt'
- move jose to devDependencies (only used in tests)
- remove redundant mock-server-capability tests that didn't exercise
  our code; share a single test DB across the describe block
