+++ backend/package.json
@@ -35,7 +35,6 @@
     "drizzle-orm": "^0.44.7",
     "elysia": "^1.4.7",
     "exa-js": "^1.0.11",
-    "jose": "6.1.2",
     "openai": "^5.22.1",
     "pino": "^9.11.0",
     "postgres": "^3.4.7",
@@ -51,6 +50,7 @@
     "drizzle-kit": "^0.31.7",
     "eslint": "^9.39.2",
     "globals": "^16.4.0",
+    "jose": "^6.1.2",
     "oauth2-mock-server": "^8.2.2",
     "pino-pretty": "^13.1.1",
     "prettier": "^3.2.4",
